requirements.txt
Flask>=0.10.1
boto>=2.13.3
ipython>=1.0.0
matplotlib>=1.3.1
numpy>=1.8.1
pandas>=0.13.1
pyyaml>=3.10
scikit_learn>=0.14.1
scipy>=0.13.3
